Move DoesNotUseOperandBuffer and CanShareOperandBufferWithUser from liveness_util to methods on TuplePointsToAnalysis and HloDataflowAnalysis.
PiperOrigin-RevId: 196903216
